% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmatrix.R
\name{cmatrix}
\alias{cmatrix}
\title{Cmatrix recoding}
\usage{
cmatrix(msa)
}
\arguments{
\item{msa}{matrix of class \code{DNAbin} or \code{AAbin}}
}
\value{
recoded Cmatrix
}
\description{
Cmatrix recoding
}
\details{
recodes gaps as increasing even and bases/acids as increasing odd integers
}
\references{
Satija R, Novak A., Mikls I., Lyngs R., and Hein J. (2009) BigFoot: Bayesian alignment and phylogenetic footprinting with MCMC, BMC Evolutionary Biology, 9, 217. Supplement page 3.

Penn et al. (2010). An alignment confidence score capturing robustness to guide tree uncertainty. Molecular Biology and Evolution. 27:1759--1767
}
