% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_derivates.R
\name{daughter_scores}
\alias{daughter_scores}
\title{Calculate additional scores based on residue pair score}
\usage{
daughter_scores(rpsc, score = "gcsc")
}
\arguments{
\item{score}{character, currently c("gcsc", "rprsc"); gcsc = guidance column score; rprsc = residue pair residue score.}

\item{polenta_obj}{usually result of \code{\link{guidance}} with score method = "Rcpp"}
}
\value{
data.frame or list of data.frames with scores
}
\description{
Calculate additional scores based on residue pair score
}
\details{
The GUIDANCE column score can be e.g. used as a sites weights in RAxML (flag -g)
}
\author{
Franz-Sebastian Krah
}
