% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposePhyloRooted.R
\name{decomposePhyloRooted}
\alias{decomposePhyloRooted}
\title{Decompose Phylogeny}
\usage{
decomposePhyloRooted(phy, k = 200)
}
\arguments{
\item{phy}{An object of class \code{\link{phylo}}.}

\item{k}{An integer giving the size of the subtrees.}
}
\description{
Decomposes (splits) a phylogenetic tree into roughly equal-sized subtree.
}
