% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMSA.R
\name{extractMSA}
\alias{extractMSA}
\title{Extract Sequences from POLENTA Output}
\usage{
extractMSA(x)
}
\arguments{
\item{x}{An object of class \code{"\link[=polentaDNA-class]{polentaDNA}"} or
\code{"\link[=polentaAA-class]{polentaAA}"}.}
}
\description{
Extract DNA or amino acid sequences from \code{polenta*}
  objects.
}
\author{
Christoph Heibl
}
