% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guidance.R
\name{guidance}
\alias{guidance}
\title{GUIDetree-based AligNment ConficencE}
\usage{
guidance(sequences, msa.exec = "/usr/local/bin/mafft", bootstrap = 100,
  ncore = 1, method = "auto", store_msas = FALSE)
}
\arguments{
\item{sequences}{object of class \code{\link{DNAbin}} or
\code{\link{AAbin}} containing unaligned sequences of DNA or amino acids.}

\item{msa.exec}{character string giving the path to the executable of the
alignment program (e.g. "/usr/local/bin/mafft"); Must be on of: 'mafft', 'muscle', 'clustalo', 'clustalw2'}

\item{bootstrap}{integer giving the number of alternative MSAs to be computed}

\item{ncore}{integer specifying the number of cores; default = 1 (serial), "auto" can be used for automated usage of all detected cores}

\item{method}{further argument passed to mafft, default is \code{"auto"}}
}
\value{
object of class \code{polenta}:

msa original MSA as computed by \code{msa.program}

scores residue pair score
}
\description{
MSA reliability assessment GUIDANCE (Penn et al. 2010)
}
\details{
Calculates column confidence (and other scors) by comparing
  alternative MSAs generated by alternative guide trees derived from
  bootstrap MSAs (Felsenstein 1985). The basic comparison between the BP MSAs
  and a reference MSA is if column residue pairs are identically aligned in
  all alternative MSAs compared with the base MSA (see \code{compareMSAs}).
}
\examples{
\dontrun{
# run GUIDANCE on example data using MAFFT
file <- system.file("extdata", "BB50009.fasta", package = "polenta")
aa_seq<- read.fas(file)
g_res <- guidance(sequences = aa_seq)
scores <- daughter_scores(g_res, score = c("gcsc", "rprsc"))
hist(scores$gcsc$score, xlab = "Column score", main = "GUIDANCE")
}

}
\references{
Felsenstein J. 1985. Confidence limits on phylogenies: an
  approach using the bootstrap. Evolution 39:783–791

Penn et al. (2010). An alignment confidence score capturing
  robustness to guide tree uncertainty. Molecular Biology and Evolution
  27:1759--1767

G. Landan and D. Graur (2008). Local reliability measures from
  sets of co-optimal multiple sequencesuence alignments. 13:15--24
}
\seealso{
\code{\link{msa_set_scoreR}}, \code{\link{guidance2}},
  \code{\link{HoT}}
}
\author{
Franz-Sebastian Krah
}
