% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guidance2.R
\name{guidance2}
\alias{guidance2}
\title{MSA reliability assessment GUIDANCE2 (Sela et al. 2015)}
\usage{
guidance2(sequences, msa.exec = "/usr/local/bin/mafft", bootstrap = 100,
  n.part = "auto", ncore = 1, method = "auto", store_msas = FALSE,
  n.coopt = "auto")
}
\arguments{
\item{sequences}{object of class \code{\link{DNAbin}} or
\code{\link{AAbin}} containing unaligned sequences of DNA or amino acids.}

\item{msa.exec}{character string giving the path to the executable of the
alignment program (e.g. "/usr/local/bin/mafft"); Must be on of: 'mafft', 'muscle', 'clustalo', 'clustalw2'}

\item{bootstrap}{integer giving the number of alternative MSAs to be computed}

\item{ncore}{integer specifying the number of cores; default = 1 (serial), "auto" can be used for automated usage of all detected cores}

\item{method}{further argument passed to mafft, default is \code{"auto"}}
}
\value{
object of class \code{polenta}:

msa original MSA as computed by \code{msa.program}

scores residue pair score
}
\description{
MSA reliability assessment GUIDANCE2 (Sela et al. 2015)
}
\details{
Calculates column confidence (and other scors) by comparing alternative MSAs generated by the GUIDANCE with varying gap opening panelty and the HoT methodology. First 100 alternative MSAs (with BP guide trees) with varying gap opening panelty are produced, then for each n (default = 4) co-optimal alignments are produced using HoT. The basic comparison between the BP MSAs and a reference MSA is if column residue pairs are identically aligned in all alternative MSAs compared with the base MSA (see \code{compareMSAs}).
}
\references{
Felsenstein J. 1985. Confidence limits on phylogenies: an
  approach using the bootstrap. Evolution 39:783–791

Penn et al. (2010). An alignment confidence score capturing
  robustness to guide tree uncertainty. Molecular Biology and Evolution
  27:1759--1767

Sela et al. (2015). GUIDANCE2: accurate detection of unreliable alignment regions accounting for the uncertainty of multiple parameters. Nucleic acids research 43:W7--W14

G. Landan and D. Graur (2008). Local reliability measures from sets of co-optimal multiple sequencesuence alignments. Pacific Symposium on Biocomputin. 13:15--24
}
\seealso{
\code{\link{compareMSAs}}, \code{\link{guidance}}, \code{\link{HoT}}
}
\author{
Franz-Sebastian Krah
}
