% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guidanceSA.R
\name{guidanceSA}
\alias{guidanceSA}
\title{Interface to guidance program}
\usage{
guidanceSA(sequences, msa.program, programm, bootstrap, gencode, outorder,
  msafile, cutoff = 0.93, moreArgs, exec, proc_num, quiet = FALSE)
}
\arguments{
\item{sequences}{An object of class \code{\link{DNAbin}} or
\code{\link{AAbin}} containing unaligned sequences of DNA or amino acids.}

\item{msa.program}{A charcter string giving the name of the MSA program,
currelty one of c("mafft", "muscle", "clustalo", "clustalw2"); MAFFT is
default}

\item{bootstrap}{An integer giving the number of perturbated MSAs.}

\item{exec}{path to guidance program folder, e.g. "/Applications/guidance.v2.02/"}

\item{proc_num}{number of cores}

\item{quiet}{logical if TRUE, progress is printed to console}

\item{parallel}{logical, if TRUE, specify the number of cores}

\item{ncore}{number of cores (default is maxinum of local threads)}

\item{program}{character, one of c("guidance", "guidance2", "HoT")}
}
\value{
list containing following scores and alignments:

mean_scores residue pair score and mean column score

column_score

residue_column_score GUIDANCE score

residue_pair_residue_score

residual_pair_sequence_pair_score

residual_pair_sequence_score

residue_pair_score

base_msa

guidance_msa is the base_MSA removed from unreliable
  residues/columns/sequences below cutoffs
}
\description{
Interface to guidance program
}
\author{
Franz-Sebastian Krah
}
