% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoT.R
\name{HoT}
\alias{HoT}
\title{Heads or Tails Alignment Reliability}
\usage{
HoT(sequences, method = "auto", bootstrap, n.coopt = "auto",
  plot_guide = TRUE, store_msas = FALSE,
  msa.exec = "/usr/local/bin/mafft", ncore = 1)
}
\arguments{
\item{sequences}{object of class \code{\link{DNAbin}} or \code{\link{AAbin}}
containing unaligned sequences of DNA or amino acids.}

\item{method}{further argument passed to MAFFT, default is \code{"auto"}}

\item{bootstrap}{integer giving the number of alternative MSAs to be computed}

\item{n.coopt}{Character string, xxx.}

\item{plot_guide}{Logical, xxx.}

\item{store_msas}{Logical, xxx.}

\item{msa.exec}{character string giving the path to the executable of the
alignment program (e.g. "/usr/local/bin/mafft"); Must be on of: 'mafft',
'muscle', 'clustalo', 'clustalw2'}

\item{ncore}{integer specifying the number of cores; default = 1 (serial),
"auto" can be used for automated usage of all detected cores.}
}
\value{
object of class \code{polenta}: msa original MSA as computed by
  \code{msa.program} scores residue pair score
}
\description{
MSA reliability assessment HoT (Landan and Graur 2008)
}
\details{
Calculates column reliability (and other scors) by comparing
  alternative MSAs generated by aligning guide tree partitions as described
  in Landan and Graur (2008). For details see \code{compareMSAs}. 8*(N-3)
  alternative MSAs are generated by default, where N is the number of
  sequences.
}
\references{
G. Landan and D. Graur (2008). Local reliability measures from
  sets of co-optimal multiple sequencesuence alignments. 13:15--24
}
\seealso{
\code{\link{msa_set_scoreR}}, \code{\link{guidance}},
  \code{\link{guidance2}}
}
\author{
Franz-Sebastian Krah
}
