% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoT_in_GUIDANCE2.R
\name{Hot_GUIDANCE2}
\alias{Hot_GUIDANCE2}
\title{XXX}
\usage{
Hot_GUIDANCE2(msa, n.coopt, type, td, files, raw_seq, msa.program, method,
  int_file, msa.exec)
}
\arguments{
\item{msa}{XXX}

\item{n.coopt}{XXX}

\item{type}{XXX}

\item{td}{XXX}

\item{files}{XXX}

\item{raw_seq}{XXX}

\item{msa.program}{XXX}

\item{method}{XXX}

\item{int_file}{XXX}

\item{msa.exec}{XXX}
}
\description{
XXX
}
