% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa_RPscore.R
\name{msa_RPscore}
\alias{msa_RPscore}
\title{computes residue pair score scores by comparing a reference MSA with alternative MSAs}
\usage{
msa_RPscore(ref, alt)
}
\arguments{
\item{ref}{reference MSA, object of class .. must be object of \code{\link{gbbin.matrix}}}

\item{alt}{either single or many alternative MSA(s). If single class \code{matrix}, if multiple class \code{list}. Must be object of \code{\link{gbbin.matrix}} or \code{\link{gbbin.list}}}
}
\description{
computes residue pair score scores by comparing a reference MSA with alternative MSAs
}
\author{
Franz-Sebastian Krah
}
