% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa_set_scoreR.R
\name{msa_set_scoreR}
\alias{msa_set_scoreR}
\title{Compare reference MSAs with alternative MSAs}
\usage{
msa_set_scoreR(ref, alt, bootstrap)
}
\arguments{
\item{ref}{of class data.frame, is the reference MSA ('BASE MSA') with
sequences as columns}

\item{alt}{path to alternative files}

\item{bootstrap}{number of alt MSAs}

\item{na.rm}{logical if gab comparisons should be deleted}
}
\value{
list containing following score:

residue_pair_score: if one alternative MSA is supplied then this is 1
  if a residue pair was identically aligned as in the reference MSA and 0
  otherwise. If more than one, then the average of the residue pair scores
  that result from each comparison with the reference MSA.
}
\description{
MSA reliability scores (Penn et al. 2010)

Rcpp code computing basic MSA comparision. The most basic is the residue pairs residue score, which checks if residue pairs combinations are correctly aligned in both MSAs.
}
\references{
Penn et al. (2010). An alignment confidence score capturing
  robustness to guide tree uncertainty. Molecular Biology and Evolution
  27:1759--1767
}
\seealso{
\code{\link{msa_set_scoreSA}}
}
\author{
Franz-Sebastian Krah
}
