% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polenta-package.R, R/polenta.R
\docType{package}
\name{polenta}
\alias{polenta}
\alias{polenta-package}
\alias{polenta}
\title{polenta: PASTA Optimal confidENce Transivitiy Alignment}
\usage{
polenta(seqs, gt, k = 200, bootstrap = 100, method = "auto", exec, ncore)
}
\arguments{
\item{seqs}{An object of class \code{\link{DNAbin}} or \code{\link{AAbin}}
containing unaligned sequences of DNA or amino acids.}

\item{gt}{\emph{Currently unused.}}

\item{k}{An integer giving the size of cluster in which the dataset is split.}

\item{bootstrap}{An integer giving the number of bootstrap replicates.}

\item{method}{A character string choosing a method of the alignment program;
see \code{\link{mafft}} for possible options.}

\item{exec}{A character string giving the path to the alignment program
executable.}

\item{ncore}{An integer giving the number of cores to use in parallel mode.}
}
\value{
An object of class \code{\link{polentaDNA}}.
}
\description{
Multiple Sequence Alignment with PASTA and GUIDANCE:
PASTA (Mirarab, Nguyen, and Warnow 2014) is a highly accurate and scalable algorithm for multiple sequence
alignment. GUIDANCE  (Penn, Privman, Landan, Grauer, and Pupko 2010) is a well-performing method for the
assessment of alignment uncertainty. Both methods are available as stand-alone programs, but,
unfortunately, cannot be combined. This package provides a reimplementation of PASTA and GUIDANCE
to combine their power to accurately align large numbers of sequences.

Provides a complete reimplementation of the PASTA algorithm
  (Mirarab, Nguyen, and Warnow 2014) in R.
}
\seealso{
\code{\link{extractMSA}} for extractiong the multiple sequence
  alignment of an \code{polentaDNA} object.
}
