% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polenta_dev.R
\name{polenta_dev}
\alias{polenta_dev}
\title{Ultra-Large Multiple Sequence Alignment with PASTA}
\usage{
polenta_dev(seqs, gt, k = 200, bootstrap = 100, method = "auto", exec,
  ncore)
}
\arguments{
\item{seqs}{An object of class \code{\link{DNAbin}} or \code{\link{AAbin}}
containing unaligned sequences of DNA or amino acids.}

\item{gt}{\emph{Currently unused.}}

\item{k}{An integer giving the size of cluster in which the dataset is split.}

\item{bootstrap}{An integer giving the number of bootstrap replicates.}

\item{method}{A character string choosing a method of the alignment program;
see \code{\link{mafft}} for possible options.}

\item{exec}{A character string giving the path to the alignment program
executable.}

\item{ncore}{An integer giving the number of cores to use in parallel mode.}
}
\value{
An object of class \code{"\link[=polentaDNA-class]{polentaDNA}"}.
}
\description{
Provides a complete reimplementation of the PASTA algorithm
  (Mirarab, Nguyen, and Warnow 2014) in R.
}
\seealso{
\code{\link{extractMSA}} for extractiong the multiple sequence
  alignment of a \code{"\link[=polentaDNA-class]{polentaDNA}"} object.
}
