% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polentaAA-class.R
\docType{class}
\name{polentaAA-class}
\alias{polentaAA-class}
\title{An S4 Class to represent Amino Acid Alignments with Quality Scores}
\description{
\code{"polentaAA"} holds a multiple sequence alignment of alino
  acids together with quality scores for each cell in the alignment.
}
\section{Slots}{

\describe{
\item{\code{msa}}{An object of class \code{\link{AAbin}}.}

\item{\code{scores}}{A matrix of quality scores.}

\item{\code{method}}{A characters string giving the method to derive the quality
scores.}
}}

\seealso{
\code{"\link[=polentaDNA-class]{polentaDNA}"}
}
