% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reappendScores_gen_dev.R
\name{reappendScores_gen_dev}
\alias{reappendScores_gen_dev}
\title{Reappend Scores after Merging}
\usage{
reappendScores_gen_dev(v, s)
}
\arguments{
\item{v}{XXX}

\item{merged}{XXX}

\item{scored}{XXX}
}
\description{
Reappend scores to alignments after profile alignment.
}
\seealso{
\code{"\link[=polentaDNA-class]{polentaDNA}"}
}
