% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitivityMerge.R
\name{transitivityMerge}
\alias{transitivityMerge}
\title{Transitivity Merge}
\usage{
transitivityMerge(x, id = c(1, 2))
}
\arguments{
\item{x}{A list containing objects of class \code{\link{polentaDNA}}.}
}
\value{
An object of class \code{\link{polentaDNA}}.
}
\description{
Merges two multiple sequence alignments and their reliability scores using the transitivity
  criterion.
}
