% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guidance.R
\name{guidance}
\alias{guidance}
\title{GUIDetree-based AligNment ConficencE}
\usage{
guidance(sequences, msa.program = "mafft",
  msa.exec = "/usr/local/bin/mafft", bootstrap = 100, parallel = FALSE,
  ncore = "auto", method = "auto", int_file = FALSE, alt.msas.file,
  nj.program = "R", score_method = "Rcpp", msa_set_score.exec)
}
\arguments{
\item{sequences}{An object of class \code{\link{DNAbin}} or
\code{\link{AAbin}} containing unaligned sequences of DNA or amino acids.}

\item{msa.program}{A charcter string giving the name of the MSA program,
currelty one of c("mafft", "muscle", "clustalo", "clustalw2"); MAFFT is
default}

\item{bootstrap}{An integer giving the number of perturbated MSAs.}

\item{parallel}{logical, if TRUE, specify the number of cores}

\item{ncore}{number of cores (default is maxinum of local threads)}

\item{method}{further arguments passed to mafft, default is \code{"auto"}}

\item{nj.program}{specify if R functions or SEMPHY should be used for guide-tree estimation}

\item{exec}{A character string giving the path to the executable of the
alignment program.}
}
\description{
MSA reliability assessment GUIDANCE (Penn et al. 2010)
}
\details{
Calculates column confidence (and other scors) by comparing
  alternative MSAs generated by alternative guide trees derived from
  bootstrap MSAs (Felsenstein 1985). The basic comparison between the BP MSAs
  and a reference MSA is if column residue pairs are identically aligned in
  all alternative MSAs compared with the base MSA (see \code{compareMSAs}).
}
\examples{
\dontrun{
# run GUIDANCE on example data using MAFFT
file <- system.file("extdata", "BB50009.fasta", package = "rpg")
aa_seq<- read.fas(file)
g_res <- guidance(sequences = aa_seq)
scores <- daughter_scores(g_r, score = c("gcsc", "rprsc"))
hist(scores$gcsc$score, xlab = "Column score", main = "GUIDANCE")
}

}
\references{
Felsenstein J. 1985. Confidence limits on phylogenies: an
  approach using the bootstrap. Evolution 39:783–791

Penn et al. (2010). An alignment confidence score capturing
  robustness to guide tree uncertainty. Molecular Biology and Evolution
  27:1759--1767

G. Landan and D. Graur (2008). Local reliability measures from
  sets of co-optimal multiple sequencesuence alignments. 13:15--24
}
\seealso{
\code{\link{compareMSAs}}, \code{\link{guidance2}},
  \code{\link{HoT}}
}
\author{
Franz-Sebastian Krah
}
