% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa_set_scoreR.R
\name{msa_set_scoreR}
\alias{msa_set_scoreR}
\title{Compare reference MSAs with alternative MSAs}
\usage{
msa_set_scoreR(ref, alt)
}
\arguments{
\item{ref}{MSA of class \code{\link{AAbin}} or \code{\link{bin}}}

\item{alt}{single MSA or list of MSAs or path to alternative files. Single MSAs and list members should be of class \code{\link{AAbin}} or \code{\link{DNAbin}}}

\item{bootstrap}{number of alt MSAs}

\item{na.rm}{logical if gab comparisons should be deleted}
}
\value{
matrix containing following scores:

residue_pair_score: if one alternative MSA is supplied then the score is 1
  if a residue pair was identically aligned as in the reference MSA and 0
  otherwise. If more than one alternative MSA is used, then the average of the residue pair scores from all comparisons
}
\description{
MSA reliability scores (Penn et al. 2010)
}
\references{
Penn et al. (2010). An alignment confidence score capturing
  robustness to guide tree uncertainty. Molecular Biology and Evolution
  27:1759--1767
}
\seealso{
\code{\link{msa_set_scoreSA}}
}
\author{
Franz-Sebastian Krah
}
