% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reappendScores.R
\name{reappendScores}
\alias{reappendScores}
\title{Reappend Scores after Merging}
\usage{
reappendScores(v, merged, scored)
}
\arguments{
\item{v}{XXX}

\item{merged}{XXX}

\item{scored}{XXX}
}
\description{
Reappend scores to alignments after profile alignment.
}
\seealso{
\code{"\link[=polentaDNA-class]{polentaDNA}"}
}
