% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPS.R
\name{RPS}
\alias{RPS}
\title{Computes residue pair score scores by comparing a two MSAs}
\usage{
RPS(ref, alt)
}
\arguments{
\item{ref}{reference MSA, object of class .. must be object of \code{\link{gbbin.matrix}}}

\item{alt}{either single or many alternative MSA(s). If single class \code{matrix}, if multiple class \code{list}. Must be object of \code{\link{gbbin.matrix}} or \code{\link{gbbin.list}}}
}
\value{
residue_pair_score if one alternative MSA is supplied then this is 1
  if a residue pair was identically aligned as in the reference MSA and 0
  otherwise. If more than one, then the average of the residue pair scores
  that result from each comparison with the reference MSA.
}
\description{
Computes residue pair score scores by comparing a two MSAs
}
\author{
Franz-Sebastian Krah
}
